new try

+++ hashmap.c
@@ -129,7 +129,8 @@ Pair * searchMap(HashMap * map,  char * key) {
 
 Pair * firstMap(HashMap * map) {
     if (map == NULL || map->size == 0)return NULL;
-    for(int k = 0 ; k < map->capacity; k++)
+
+    for(long k = 0 ; k < map->capacity; k++)
     {
         if (map->buckets[k] != NULL && map->buckets[k]->key != NULL)
         {
